Wait when trying to play a webstream while connecting to wifi
The previous code had some issues. First of all, with
DONT_ACCEPT_SAME_RFID_TWICE it would not work at all. But without a
change, not even that works correctly (due to gTriedToConnectToHost
never being set).

IMHO the code is simplified a lot. The whole control flow can be
simplified if we just wait until a connection try hast completed.